+++ lib/functions/load-assets.php
@@ -26,7 +26,7 @@ function enqueue_assets() {
 	wp_enqueue_script( CHILD_TEXT_DOMAIN . '-responsive-menu', CHILD_URL . '/assets/js/dist/responsive-menu.min.js', array( 'jquery' ), CHILD_THEME_VERSION, true );
 
 
-	wp_enqueue_script( CHILD_TEXT_DOMAIN . '-nav-scroll', CHILD_URL . '/assets/js/dist/nav-scroll.min.js', array( 'jquery' ), CHILD_THEME_VERSION, true );
+	//wp_enqueue_script( CHILD_TEXT_DOMAIN . '-nav-scroll', CHILD_URL . '/assets/js/dist/nav-scroll.min.js', array( 'jquery' ), CHILD_THEME_VERSION, true );
 
 
 	$localized_script_args = array(